pr comments rename tests
