+++ e2e/e2etests/test_v2_eth_withdraw_and_arbitrary_call.go
@@ -1,0 +1,40 @@
+package e2etests
+
+import (
+	"math/big"
+
+	"github.com/stretchr/testify/require"
+	"github.com/zeta-chain/protocol-contracts/v2/pkg/gatewayzevm.sol"
+
+	"github.com/zeta-chain/node/e2e/runner"
+	"github.com/zeta-chain/node/e2e/utils"
+	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
+)
+
+const payloadMessageWithdrawETH = "this is a test ETH withdraw and call payload"
+
+func TestV2ETHWithdrawAndArbitraryCall(r *runner.E2ERunner, args []string) {
+	require.Len(r, args, 1)
+
+	amount, ok := big.NewInt(0).SetString(args[0], 10)
+	require.True(r, ok, "Invalid amount specified for TestV2ETHWithdrawAndCall")
+
+	r.AssertTestDAppEVMCalled(false, payloadMessageWithdrawETH, amount)
+
+	r.ApproveETHZRC20(r.GatewayZEVMAddr)
+
+	// perform the withdraw
+	tx := r.V2ETHWithdrawAndCall(
+		r.TestDAppV2EVMAddr,
+		amount,
+		r.EncodeGasCall(payloadMessageWithdrawETH),
+		gatewayzevm.RevertOptions{OnRevertGasLimit: big.NewInt(0)},
+	)
+
+	// wait for the cctx to be mined
+	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
+	r.Logger.CCTX(*cctx, "withdraw")
+	require.Equal(r, crosschaintypes.CctxStatus_OutboundMined, cctx.CctxStatus.Status)
+
+	r.AssertTestDAppEVMCalled(true, payloadMessageWithdrawETH, amount)
+}
